% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlations}
\alias{correlations}
\title{Create Correlation Matrix}
\usage{
correlations(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
prints a correlation matrix
}
\description{
Produces a correlation matrix of all numeric variables in a dataframe
}
\details{
The function will select only the numeric columns to include
}
\examples{
correlations(df)
}
\author{
Valery Lynn
}
