% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_describe.R
\name{data_describe}
\alias{data_describe}
\title{data description}
\usage{
data_describe(df)
}
\arguments{
\item{a}{dataframe with some numerical variables}
}
\value{
a printout of the dataframe shape and a list of column names
}
\description{
Prints out the dataframe shape and the column names
}
\examples{
data_describe(df)
}
\author{
Valery Lynn
}
