% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stats.R
\name{data_stats}
\alias{data_stats}
\title{Descriptive statistics}
\usage{
data_stats(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
prints a summary of descriptive statistics
}
\description{
Prints out descriptive statistics for the variables in a dataframe
}
\examples{
data_stats(df)
}
\author{
Valery Lynn
}
