% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\name{data_types}
\alias{data_types}
\title{data types}
\usage{
data_types(df)
}
\arguments{
\item{df}{a dataframe with some numerical columns}
}
\value{
a printout of the dataframe data types for all variables
}
\description{
Prints out the data types for the variables in a dataframe
}
\examples{
data_types(df)
}
\author{
Valery Lynn
}
