% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dens_plots.R
\name{dens_plots}
\alias{dens_plots}
\title{Density plots}
\usage{
dens_plots(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
prints a kernel density plot for every numeric variable
}
\description{
Creates kernel density plots for all numeric variables in a dataframe
}
\details{
The function will select only the numeric columns to include
}
\examples{
dens_plots(df)
}
\author{
Valery Lynn
}
