% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_df.R
\name{make_df}
\alias{make_df}
\title{Make a numeric dataframe}
\usage{
make_df(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
returns a dataframe with only numeric columns
}
\description{
Selects only the numeric columns in a dataframe
}
\details{
The function will select only the numeric columns to include
}
\examples{
make_df(df)
}
\author{
Valery Lynn
}
