% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plots.R
\name{qq_plots}
\alias{qq_plots}
\title{Print QQ-plots}
\usage{
qq_plots(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
returns the output of the step function (anova)
}
\description{
Prints QQ-plots for all variables in a dataframe
}
\details{
The function will select only the numeric columns to include
}
\examples{
qq_plots(df)
}
\author{
Valery Lynn
}
