% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_eda.R
\name{quick_eda}
\alias{quick_eda}
\title{Quick EDA}
\usage{
quick_eda(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
returns an exploratory data analysis for all variables in a dataframe
}
\description{
Makes a quick eda by running all the functions in the quickeda package
Returns data description, data types, density plots, qq-plots, correlation matrix,
heatmap, and the anova summary of a linear regression run with the step function
}
\details{
The function will select only the numeric columns to include where needed
}
\examples{
quick_eda(df)
}
\author{
Valery Lynn
}
